DRIP-20 checkout created now fires after customer login, checkout update no longer fires everytime cart page is loaded

+++ app/code/community/Drip/Connect/Helper/Data.php
@@ -118,6 +118,6 @@ class Drip_Connect_Helper_Data extends Mage_Core_Helper_Abstract
      * @return string
      */
     public function priceAsCents($price) {
-        return preg_replace("/[^0-9.]/", "", $price) * 100;
+        return (int) preg_replace("/[^0-9.]/", "", $price) * 100;
     }
 }
